# authservice/openapi.yaml
openapi: 3.0.2
info:
  title: rokwire-auth
  description: rokwire auth
  version: "0.1.0alpha"
paths:
  /authservice/verifications:
    post:
      summary: initiate a verification
      operationId: auth_rest_service.initiate_verification
      responses:
        201:
          description: initiated
        400:
          description: bad request
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationInitBody'
  /authservice/verification-check:
    post:
      summary: verification check
      operationId: auth_rest_service.verification_check
      responses:
        200:
          description: verification result
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                        enum:
                          - true
                      id_token:
                        type: string
                    required:
                      - success
                      - id_token
                  - type: object
                    properties:
                      success:
                        type: boolean
                        enum:
                          - false
                    required:
                      - success
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationCheckBody'
  /authservice/ping:
    get:
      summary: ping
      operationId: auth_rest_service.ping
      responses:
        200:
          description: success
components:
  schemas:
    PhoneNumber:
      type: string
      minLength: 10
      maxLength: 10
      # pattern: '^\d{10}$'
    VerificationInitBody:
      type: object
      properties:
        phoneNumber:
          $ref: '#/components/schemas/PhoneNumber'
        channel:
          type: string
          enum:
            - sms
            - call
      required:
        - phoneNumber
        - channel
      example:
        phoneNumber: "1234567890"
        channel: "sms"
    VerificationCheckBody:
      type: object
      properties:
        phoneNumber:
          $ref: '#/components/schemas/PhoneNumber'
        code:
          type: string
          minLength: 6
          maxLength: 6
      required:
        - phoneNumber
        - code
      example:
        phoneNumber: "1234567890"
        code: "123456"